+++ src/i18n/index.js
@@ -10,6 +10,7 @@ import de from './de';
 import da from './da';
 import zh_tw from './zh_tw';
 import pl from './pl';
+import es from './es';
 import ja from './ja';
 
 module.exports = {
@@ -25,5 +26,6 @@ module.exports = {
   da,
   zh_tw,
   pl,
+  es,
   ja
 };